[master] Moves user repository
