+++ app/src/main/java/de/cashflow/cashflow/user/repository/UserRepository.kt
@@ -1,4 +1,4 @@
-package de.cashflow.cashflow.login.repository
+package de.cashflow.cashflow.user.repository
 
 interface UserRepository {
